requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
WTForms==3.0.1
email-validator==2.0.0
Werkzeug==2.3.7
Flask-Talisman==1.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
